refactor: move Guidance under Runner namespace
Reviewer::Guidance → Reviewer::Runner::Guidance since it is only
used by Runner. Delete old lib/reviewer/guidance.rb.
