+++ lib/reviewer.rb
@@ -16,7 +16,6 @@ require_relative 'reviewer/batch'
 require_relative 'reviewer/capabilities'
 require_relative 'reviewer/command'
 require_relative 'reviewer/failed_files'
-require_relative 'reviewer/guidance'
 require_relative 'reviewer/keywords'
 require_relative 'reviewer/output'
 require_relative 'reviewer/prompt'
